created host_table.go file and fixed CreateEvent func
